Refactor Flask Application for Improved Structure and Maintainability
Major Changes:

Application Factory Pattern Implemented:

Created an application factory function create_app() in main.py to initialize the Flask app. This method improves the application's structure and makes it easier to manage different configurations and testing environments.
Configurations are now loaded from src/config.py using app.config.from_object(Config) to centralize and simplify configuration management.
Database Initialization Refactored:

Extracted the SQLAlchemy database instance db to a separate file database.py to avoid circular dependencies and centralize database interactions.
Updated import statements across the application, particularly in models.py, to import db from database.py instead of main.py.
User Loading Function for Flask-Login:

Added a load_user function within create_app() in main.py to integrate with Flask-Login for user session management. This function is necessary for Flask-Login to load user information.
Blueprints Preparation:

Prepared the application for the future integration of Flask Blueprints by organizing views and routes. This step sets the foundation for modularizing the application into distinct components.
Testing Configuration Adjusted:

Modified test scripts to align with the new application structure. Ensured that tests run correctly with the updated application factory and database initialization methods.
Miscellaneous Code Enhancements:

Performed minor code cleanups, added comments for better readability, and improved overall code organization.
Notes:

These changes were made to address a circular dependency issue and to professionalize the application structure.
The application's core functionality remains unchanged, but the structure is now more scalable and easier to maintain.
