+++ run.py
@@ -1,0 +1,6 @@
+from main import create_app
+
+app = create_app()
+
+if __name__ == "__main__":
+	app.run()
